requirements.txt
# Filename: requirements.txt

streamlit
google-generativeai
